+++ HellDivers_UnityProject/Assets/Resources/Mobs/Effect/BloodGushFX.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 7ae7825c17250f349bf1aef379d68d87
-NativeFormatImporter:
+guid: 6a73634de3adf9b4096a6d4f7c8a153c
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 0
   userData: 
   assetBundleName: 
   assetBundleVariant:
