+++ HellDivers_UnityProject/Assets/Resources/Mobs/Effect/BloodGushFX/BloodPartcleFX_Scene.unity.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cb2d6e429ca112f499e1db771d5bd27c
+timeCreated: 1476032555
+licenseType: Store
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
