+++ HellDivers_UnityProject/Assets/Resources/Mobs/Effect/BloodGushFX/BloodSprayFX.prefab.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 7ae7825c17250f349bf1aef379d68d87
+guid: d32aa916f50cac94db11e6f19821a58f
+timeCreated: 1476036105
+licenseType: Store
 NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
   userData: 
   assetBundleName: 
   assetBundleVariant:
