+++ HellDivers_UnityProject/Assets/Resources/Mobs/Effect/BloodGushFX/Materials.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: b6306e0edb49a9d448cd3a806e7684a6
+folderAsset: yes
+timeCreated: 1476032489
+licenseType: Store
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
