+++ HellDivers_UnityProject/Assets/Resources/Mobs/Effect/BloodGushFX/Materials/BloodParticleA.mat.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 7ae7825c17250f349bf1aef379d68d87
+guid: 94a55713ee628a245a74794f36a1b6f7
+timeCreated: 1476029978
+licenseType: Store
 NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
   userData: 
   assetBundleName: 
   assetBundleVariant:
