+++ HellDivers_UnityProject/Assets/Resources/Mobs/Effect/BloodGushFX/Materials/BloodParticleB.mat.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 7ae7825c17250f349bf1aef379d68d87
+guid: 372655694a62c8f4cba7e4a00bda0cc8
+timeCreated: 1476029978
+licenseType: Store
 NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
   userData: 
   assetBundleName: 
   assetBundleVariant:
