+++ HellDivers_UnityProject/Assets/Resources/Mobs/Effect/BloodGushFX/Materials/BloodParticleC.mat.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 7ae7825c17250f349bf1aef379d68d87
+guid: 1226a3e53649e09428c4c0cf03714527
+timeCreated: 1476029978
+licenseType: Store
 NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
   userData: 
   assetBundleName: 
   assetBundleVariant:
