+++ HellDivers_UnityProject/Assets/Resources/Mobs/Effect/BloodGushFX/Textures.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 6f3da53cc0c454f4989aeff372634d6c
+folderAsset: yes
+timeCreated: 1476032482
+licenseType: Store
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
