+++ HellDivers_UnityProject/Assets/Resources/UI/InGame/Map/PlayerPoint.prefab
@@ -48,7 +48,7 @@ MonoBehaviour:
       m_Calls: []
     m_TypeName: UnityEngine.UI.MaskableGraphic+CullStateChangedEvent, UnityEngine.UI,
       Version=1.0.0.0, Culture=neutral, PublicKeyToken=null
-  m_Sprite: {fileID: 21300000, guid: 1e555e1bc111bd24db51bff91091b9d7, type: 3}
+  m_Sprite: {fileID: 21300078, guid: 5df2a2393fdbcc0469cdb37d485a2f82, type: 3}
   m_Type: 0
   m_PreserveAspect: 0
   m_FillCenter: 1
