--- HellDivers_UnityProject/Assets/Test/Woody/EnemyAI/AITest/RandomTest.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: eed944dc14e7b9b4ab00c7811ea9be48
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
