--- HellDivers_UnityProject/Assets/Test/Woody/MiniMapCamera.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1e7149e5929714849879cc14e66abeda
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
