--- HellDivers_UnityProject/Assets/Test/Woody/RadarTest/Radar.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8b857639db1281446a48394d270a1a99
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
